+++ examples/3qubit_gates.jl
@@ -36,9 +36,11 @@ function test_toffoli()
 
     "elementary_gates" => ["T1", "T2", "T3", "H3", "cnot_12", "cnot_13", "cnot_23", "Tdagger2", "Tdagger3", "Identity"], 
     "target_gate" => QCO.ToffoliGate(),
+    "input_circuit" => toffoli_circuit(),
     
     "objective" => "minimize_depth", 
     "decomposition_type" => "exact",
+    "relax_integrality" => true,
     
     "optimizer" => "cplex",
     
@@ -46,4 +48,24 @@ function test_toffoli()
 
     return params
     
+end
+
+function toffoli_circuit()
+    # [(depth, gate)]
+    return [(1, "H3"), 
+        (2, "cnot_23"), 
+        (3, "Tdagger3"), 
+        (4, "cnot_13"), 
+        (5, "T3"), 
+        (6, "cnot_23"), 
+        (7, "Tdagger3"), 
+        (8, "cnot_13"), 
+        (9, "T2"), 
+        (10, "T3"), 
+        (11, "cnot_12"), 
+        (12, "H3"), 
+        (13, "T1"), 
+        (14, "Tdagger2"),
+        (15, "cnot_12")
+        ] 
 end